// backend/src/modules/operations/placements/operations-placements.types.ts
export type PlacementJoiningStatus =
  | "joining_pending"
  | "joined"
  | "did_not_join";